pagebrief-backend-pro/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.4
celery==5.5.3
redis==6.4.0
httpx==0.28.1
pypdf==6.0.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic-settings==2.10.1
email-validator==2.2.0
orjson==3.11.2
pytest==8.4.1
